factored out base class into separate module
